+++ tests/testthat/test_get_na_counts.R
@@ -2,6 +2,11 @@ test_that(desc="get_na_counts works as expected",
                     code={
 
                       skip_on_oldrel()
+expect_error(get_na_counts(as.POSIXct(Sys.Date())),
+             
+"get_na_counts not available for objects of class POSIXct. Try 'na_counts(x)' instead.",
+fixed=TRUE)               
+                               
                   expect_equal(nrow(get_na_counts(iris,"Species")),
                                          3)
                   expect_error(get_na_counts(airquality,
@@ -11,5 +16,11 @@ test_that(desc="get_na_counts works as expected",
                   expect_equal(nrow(get_na_counts(airquality,
                                                             grouping_cols = "Month")),
                                          5)
+                  
+                  # If using na_counts, expect that we get the same value as 
+                  # when calling via get_na_counts, using Ozone as an example 
+                  # here. 
+                  expect_equal(get_na_counts(airquality)[[1]],
+                               na_counts(airquality$Ozone))
 
                     })